move validator to pkg
