+++ cmd/server/accounts/router.go
@@ -13,15 +13,13 @@ import (
 
 	"github.com/moov-io/ach"
 	moovhttp "github.com/moov-io/base/http"
-
-	"github.com/moov-io/customers/cmd/server/accounts/validator"
-
 	"github.com/moov-io/customers/cmd/server/fed"
 	"github.com/moov-io/customers/cmd/server/route"
 	"github.com/moov-io/customers/pkg/client"
 	"github.com/moov-io/customers/pkg/secrets"
 	"github.com/moov-io/customers/pkg/secrets/hash"
 	"github.com/moov-io/customers/pkg/secrets/mask"
+	"github.com/moov-io/customers/pkg/validator"
 
 	"github.com/go-kit/kit/log"
 	"github.com/gorilla/mux"
