+++ cmd/server/main.go
@@ -21,16 +21,16 @@ import (
 	"github.com/moov-io/base/http/bind"
 	"github.com/moov-io/customers"
 	"github.com/moov-io/customers/cmd/server/accounts"
-	"github.com/moov-io/customers/cmd/server/accounts/validator"
-	"github.com/moov-io/customers/cmd/server/accounts/validator/microdeposits"
-	"github.com/moov-io/customers/cmd/server/accounts/validator/mx"
-	"github.com/moov-io/customers/cmd/server/accounts/validator/plaid"
 	"github.com/moov-io/customers/cmd/server/fed"
 	"github.com/moov-io/customers/cmd/server/paygate"
 	"github.com/moov-io/customers/internal/database"
 	"github.com/moov-io/customers/internal/util"
 	"github.com/moov-io/customers/pkg/configuration"
 	"github.com/moov-io/customers/pkg/secrets"
+	"github.com/moov-io/customers/pkg/validator"
+	"github.com/moov-io/customers/pkg/validator/microdeposits"
+	"github.com/moov-io/customers/pkg/validator/mx"
+	"github.com/moov-io/customers/pkg/validator/plaid"
 
 	"github.com/go-kit/kit/log"
 	"github.com/gorilla/mux"
