+++ pkg/validator/plaid/strategy.go
@@ -9,8 +9,8 @@ import (
 	"net/http"
 
 	"github.com/mitchellh/mapstructure"
-	"github.com/moov-io/customers/cmd/server/accounts/validator"
 	"github.com/moov-io/customers/pkg/client"
+	"github.com/moov-io/customers/pkg/validator"
 	"github.com/plaid/plaid-go/plaid"
 )
