+++ pkg/validator/plaid/strategy_test.go
@@ -11,8 +11,8 @@ import (
 	"testing"
 
 	"github.com/moov-io/base"
-	"github.com/moov-io/customers/cmd/server/accounts/validator"
 	"github.com/moov-io/customers/pkg/client"
+	"github.com/moov-io/customers/pkg/validator"
 	"github.com/plaid/plaid-go/plaid"
 	"github.com/stretchr/testify/require"
 )
