+++ pkg/validator/testvalidator/strategy_test.go
@@ -7,8 +7,8 @@ package testvalidator
 import (
 	"testing"
 
-	"github.com/moov-io/customers/cmd/server/accounts/validator"
 	"github.com/moov-io/customers/pkg/client"
+	"github.com/moov-io/customers/pkg/validator"
 	"github.com/stretchr/testify/require"
 )
